fix data type

+++ src/controller/activate.py
@@ -88,7 +88,7 @@ def img_preproc(in_image, config=None):
 
     # cut useless content
     in_image = in_image[frame_cut[0][0]:-frame_cut[0][1],
-                        frame_cut[1][0]:-frame_cut[1][1], :]
+                        frame_cut[1][0]:-frame_cut[1][1], :].astype("float32")
 
     # re-normalize image
     if mode in [0, 1]:
